move keras model builder together with keras base model
