+++ pyzoo/test/zoo/automl/model/test_base_keras_model.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 from unittest import TestCase
-from zoo.automl.model import KerasModelBuilder
+from zoo.automl.model.base_keras_model import KerasModelBuilder
 import numpy as np
 import tensorflow as tf
 import pytest
