+++ pyzoo/zoo/examples/automl/plugin_model_demo.py
@@ -16,7 +16,7 @@
 #
 
 from zoo.automl.search import SearchEngineFactory
-from zoo.automl.model import KerasModelBuilder
+from zoo.automl.model.base_keras_model import KerasModelBuilder
 from zoo.automl.model.base_pytorch_model import PytorchModelBuilder
 import torch
 import tensorflow as tf
